--- dot_config/nvim/lua/modules/hydra/init.lua
@@ -1,1 +1,0 @@
-require("modules.hydra.window")
