+++ dot_config/nvim/lua/modules/language/servers.lua.tmpl
@@ -55,16 +55,6 @@ html = {},
 cssls = {},
 {{- $servers = add1 $servers -}}
 {{- end }}
-{{ if lookPath "vscode-json-language-server" -}}
-jsonls = {
-  settings = {
-    json = {
-      schemas = require("schemastore").json.schemas(),
-    },
-  },
-},
-{{- $servers = add1 $servers -}}
-{{- end }}
 {{ if lookPath "vscode-eslint-language-server" -}}
 eslint = {},
 {{- $servers = add1 $servers -}}
@@ -161,6 +151,26 @@ vim.api.nvim_create_autocmd("LspAttach", {
     vim.keymap.set("n", "gi", vim.lsp.buf.implementation, { buffer = true })
   end
 })
+
+{{ if lookPath "vscode-json-language-server" -}}
+vim.api.nvim_create_autocmd("FileType", {
+  pattern = { "json", "jsonc" },
+  group = vim.api.nvim_create_augroup("jsonls_lazy", { clear = true }),
+  desc = "Enable jsonls with schemas on JSON file open",
+  once = true,
+  callback = function()
+    local schemas = require("schemastore").json.schemas()
+    vim.lsp.config("jsonls", {
+      settings = {
+        json = {
+          schemas = schemas,
+        },
+      },
+    })
+    vim.lsp.enable("jsonls")
+  end,
+})
+{{- end }}
 
 function M.setup()
   {{ if gt $servers 0 -}}
